requirements.txt
setuptools==73.0.1
numpy==1.24.4
pillow==10.3.0
torchvision==0.19.1
pandas==2.0.3
scikit-learn==1.3.2